LC-koko-eating-bananas [Runtime: N/A; Memory: N/A]

+++ binary-search/koko-eating-bananas.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int countTotalHour(vector<int>& piles,int speed){
+    long long countTotalHour(vector<int>& piles,int speed){
         long long totalH=0;
         for(int i: piles){
             totalH += ceil((double)i/speed);
